raspi/requirements.txt
mysql-connector-python==8.3.0
enviroplus==0.0.6
pms5003==0.0.5
pimoroni-bme280==0.0.2